### 2023-07-22 #5676 @editorialbot
[PRE REVIEW]: measr: Bayesian psychometric measurement using Stan
**Submitting author:** @wjakethompson (<a href="http://orcid.org/0000-0001-7339-0300">W. Jake Thompson</a>)
**Repository:** https://github.com/wjakethompson/measr
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.3.1
