### 2023-08-04 #5676 @smasongarrison
@chartgerink I'm not a Bayesian, but I can evaluate the measurement side of the package.
