### 2023-08-05 #5676 @Wenchao-Ma
@chartgerink<https://github.com/chartgerink> I can help with the modeling part, but not Bayesian too.

From: Mason Garrison ***@***.***>
Sent: Friday, August 4, 2023 10:57 AM
Cc: Wenchao Ma ***@***.***>; Mention ***@***.***>
Subject: [EXTERNAL] Re: [openjournals/joss-reviews] [PRE REVIEW]: measr: Bayesian psychometric measurement using Stan (Issue #5676)

@chartgerink<https://github.com/chartgerink> I'm not a Bayesian, but I can evaluate the measurement side of the package.

-
You are receiving this because you were mentioned.Message ID: ***@***.******@***.***>>
