### 2023-08-23 dac2d21
start model evaluation vignette
