+++ vignettes/articles/model-evaluation.Rmd
@@ -1,0 +1,230 @@
+---
+title: "Evaluating diagnostic classification models"
+output: rmarkdown::html_vignette
+bibliography: bib/references.bib
+csl: bib/apa.csl
+link-citations: true
+vignette: >
+  %\VignetteIndexEntry{Evaluating diagnostic classification models}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+library(wjake)
+library(showtext)
+library(glue)
+
+set_theme(base_family = "Open Sans",
+          plot_margin = ggplot2::margin(10, 10, 10, 10))
+
+font_add_google("Open Sans")
+showtext_auto()
+showtext_opts(dpi = 192)
+
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 7.2916667,
+  fig.align = "center",
+  out.width = "90%"
+)
+
+options(mc.cores = 4,
+        tidyverse.quiet = TRUE)
+```
+
+In this article, we will describe the different options for evaluating diagnostic classification models (DCMs; also known as cognitive diagnostic models [CDMs]) using measr.
+We start with the data to analyze, estimate our DCM, and learn how to evaluate different aspects of the model such as model fit and reliability.
+
+To use the code in this article, you will need to install and load the measr package.
+
+```{r}
+library(measr)
+```
+
+## Example Data
+
+To demonstrate the model fit functionality of measr, we'll examine a simulated data set.
+This data set contains 1,000 respondents and 20 items that measure a total of 4 attributes, but no item measures more than 2 attributes.
+The data was generated from the loglinear cognitive diagnostic model (LCDM), which is a general model that subsumes many other DCM subtypes [@lcdm].
+To demonstrate model fit functionality, we'll first fit an LCDM and a deterministic-input, noisy "and" gate (DINA) model [@dina] to the data set in order to compare the fit indices.
+Because the LCDM was used to generate our fake data, we expect our estimated LCDM model to perform well.
+On the other hand, the DINA model places heavy constraints on the LCDM, and therefore we expect worse performance from the DINA model.
+For details on model estimation, see [Estimating diagnostic classification models](model-estimation.html).
+
+```{r read-data, eval = FALSE, include = FALSE}
+library(tidyverse)
+
+dat <- read_rds("data/example-data.rds")
+sim_data <- dat$data[, -1]
+sim_qmatrix <- dat$q_matrix[, -1]
+
+write_rds(list(data = sim_data, qmatrix = sim_qmatrix),
+          file = "vignettes/articles/data/sim-data.rds")
+```
+
+```{r estimate-models}
+library(tidyverse)
+
+sim_data <- read_rds("data/sim-data.rds")
+
+lcdm <- measr_dcm(data = sim_data$data, qmatrix = sim_data$qmatrix,
+                  type = "lcdm", backend = "cmdstanr",
+                  iter_warmup = 1000, iter_sampling = 500,
+                  chains = 4, parallel_chains = 4, file = "fits/lcdm.rds")
+
+dina <- measr_dcm(data = sim_data$data, qmatrix = sim_data$qmatrix,
+                  type = "dina", backend = "cmdstanr",
+                  iter_warmup = 1000, iter_sampling = 500,
+                  chains = 4, parallel_chains = 4, file = "fits/dina.rds")
+```
+
+## Model Evaluation
+
+There are three major types of evaluations that are supported by measr.
+
+* Absolute model fit
+* Relative model fit (i.e., model comparisons)
+* Reliability
+
+Each of these are discussed in turn.
+
+### Absolute Model Fit
+
[…]
