### 2023-08-24 eabe3c9
update installation instructions (#23)
