### 2023-08-24 691de77
add cmdstanr to pkgdown
