+++ .github/workflows/pkgdown.yaml
@@ -33,8 +33,16 @@ jobs:
           extra-packages: |
             any::pkgdown
             any::tidyverse
+            github::stan-dev/cmdstanr
             local::.
+          upgrade: 'TRUE'
           needs: website
+
+      - name: Install cmdstan
+        run: |
+          cmdstanr::check_cmdstan_toolchain(fix = TRUE)
+          cmdstanr::install_cmdstan(cores = 2)
+        shell: Rscript {0}
 
       - name: Build site
         run: pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
