### 2023-08-24 a288b39
add force functionality to model evaluation functions
