+++ R/reliability.R
@@ -1,6 +1,9 @@
 #' Estimate the reliability of psychometric models
 #'
 #' @param model The estimated model to be evaluated.
+#' @param force If reliability information has already been added to the model
+#'   object with [add_reliability()], should it be recalculated. Default is
+#'   `FALSE`.
 #'
 #' @export
 reliability <- function(model) {
@@ -62,7 +65,12 @@ reliability <- function(model) {
 #' )
 #'
 #' reliability(rstn_mdm_lcdm)
-reliability.measrdcm <- function(model) {
+reliability.measrdcm <- function(model, force = FALSE) {
+  if ((!is.null(model$reliability) && length(model$reliability) > 0) &&
+      !force) {
+    return(model$reliability)
+  }
+
   # coerce model into a list of values required for reliability
   obj <- reli_list(model)
   att_names <- model$data$qmatrix %>%
