+++ man/reliability.Rd
@@ -7,10 +7,14 @@
 \usage{
 reliability(model)
 
-\method{reliability}{measrdcm}(model)
+\method{reliability}{measrdcm}(model, force = FALSE)
 }
 \arguments{
 \item{model}{The estimated model to be evaluated.}
+
+\item{force}{If reliability information has already been added to the model
+object with \code{\link[=add_reliability]{add_reliability()}}, should it be recalculated. Default is
+\code{FALSE}.}
 }
 \value{
 For class \code{measrdcm}, a list with 3 elements:
