+++ man/fit_m2.Rd
@@ -5,14 +5,18 @@
 \title{Estimate the \ifelse{html}{\out{M<sub>2</sub>}}{\eqn{M_2}} fit statistic for
 diagnostic classification models}
 \usage{
-\method{fit_m2}{measrdcm}(model, ci = 0.9, ...)
+\method{fit_m2}{measrdcm}(model, ..., ci = 0.9, force = FALSE)
 }
 \arguments{
 \item{model}{An estimated diagnostic classification model.}
 
+\item{...}{Unused, for extensibility.}
+
 \item{ci}{The confidence interval for the RMSEA.}
 
-\item{...}{Unused, for extensibility.}
+\item{force}{If the \ifelse{html}{\out{M<sub>2</sub>}}{\eqn{M_2}} has already
+been saved to the model object with \code{\link[=add_fit]{add_fit()}}, should it be recalculated.
+Default is \code{FALSE}.}
 }
 \value{
 A data frame created by \code{\link[dcm2:fit_m2]{dcm2::fit_m2()}}.
