### 2023-08-24 512be12
fix warning for generics
