### 2023-08-24 6567839
remove cmdstanr for pkgdown
