### 2023-08-24 25c4126
add section on saving output
