### 2023-09-20 73bc3bd
update intro vignettes to use simulated data
