+++ vignettes/articles/model-estimation.Rmd
@@ -38,25 +38,25 @@ We start with the data to analyze, understand how to specify a DCM to estimate, 
 
 To use the code in this article, you will need to install and load the measr package.
 
-```{r}
+```{r load-measr}
 library(measr)
 ```
 
-## Examination for the Certificate of Proficiency in English Data
+## Example Data
 
-To demonstrate how measr works, we will examine data from the grammar section of the Examination for the Certificate of Proficiency in English (ECPE).
-This section of the ECPE consists of 28 item multiple-choice items that together measure three attributes: morphosyntactic rules, cohesive rules, and lexical rules.
-In total, data is available for 2,922 respondents.
-This data set has been widely studied in the DCM literature [e.g., @chen2018; @liu2019; @hdcm].
-Additionally, parameter estimates from other software (Mplus) are readily available from @templin2013a, so we can compare the parameter estimates from measr to estimates derived from other software.
+To demonstrate the model estimation functionality of measr, we'll examine a simulated data set.
+This data set contains 2,000 respondents and 20 items that measure a total of 4 attributes, but no item measures more than 2 attributes.
+The data was generated from the loglinear cognitive diagnostic model (LCDM), which is a general model that subsumes many other DCM subtypes [@lcdm].
+By using a simulated data set, we can compare the parameter estimates from measr to the true data generating parameters.
 
-The ECPE data is included within the measr package.
-For additional information on these data, see `?ecpe_data`.
+```{r explore-data}
+library(tidyverse)
 
-```{r}
-ecpe_data
+sim_data <- read_rds("data/simulated-data.rds")
 
-ecpe_qmatrix
+sim_data$data
+
+sim_data$q_matrix
 ```
 
 ## Specifying a DCM for Estimation
@@ -86,34 +86,34 @@ If a file is specified, an R object of the fitted model will be automatically sa
 If the specified `file` already exists, then the fitted model will be read back into R, eliminating the need to re-estimate the model.
 
 ```{r}
-ecpe <- measr_dcm(data = ecpe_data, qmatrix = ecpe_qmatrix,
-                  resp_id = "resp_id", item_id = "item_id", type = "lcdm",
-                  method = "mcmc", backend = "cmdstanr",
-                  chains = 4, iter_warmup = 1000, iter_sampling = 500,
-                  parallel_chains = 4,
-                  file = "fits/ecpe-lcdm")
+lcdm <- measr_dcm(data = sim_data$data, qmatrix = sim_data$q_matrix,
+                  resp_id = "resp_id", item_id = "item_id",
+                  type = "lcdm", method = "mcmc", backend = "cmdstanr",
+                  iter_warmup = 1000, iter_sampling = 500,
+                  chains = 4, parallel_chains = 4,
+                  file = "fits/sim-lcdm")
 ```
 
 ### Examining Parameter Estimates
 
-Now that we've estimated a model, let's compare our parameter estimates to those reported for the ECPE data by @templin2013a.
+Now that we've estimated a model, let's compare our parameter estimates to the true values used to generate the data.
 We can start be looking at our estimates using `measr_extract()`.
 This function extracts different aspects of a model estimated with measr.
 Here, the `estimate` column reports estimated value for each parameter and a measure of the associated error (i.e., the standard deviation of the posterior distribution).
-For example, item E1 has four parameters, as it measures two attributes:
+For example, item A1 measures two attributes and therefore has four parameters:
 
-1. An intercept, which represents the log-odds of providing a correct response for a respondent who is proficient in neither of the attributes this item measures (i.e., morphosyntactic rules and cohesive rules).
[…]
