+++ vignettes/articles/model-estimation.Rmd
@@ -159,16 +159,16 @@ Again, we see relatively strong agreement between the estimates from our model a
 #|   Figure shows a strong correlation between item parameters, with only a few
 #|   discrepancies off of the line of perfect agreement.
 
-true_strc <- create_profiles(4) %>% 
-  rowwise() %>% 
-  mutate(label = glue("[{att1},{att2},{att3},{att4}]")) %>% 
-  ungroup() %>% 
-  select(label) %>% 
+true_strc <- create_profiles(4) %>%
+  rowwise() %>%
+  mutate(label = glue("[{att1},{att2},{att3},{att4}]")) %>%
+  ungroup() %>%
+  select(label) %>%
   mutate(value = sim_data$true_strc)
 
-strc_compare <- measr_extract(lcdm, "strc_param") %>% 
+strc_compare <- measr_extract(lcdm, "strc_param") %>%
   left_join(true_strc, join_by(class == label),
-            relationship = "one-to-one") %>% 
+            relationship = "one-to-one") %>%
   mutate(measr_est = map_dbl(estimate, mean))
 
 strc_compare %>%
