### 2023-09-20 7ab26b0
update deprecated stan syntax
