+++ R/stan-gqs.R
@@ -2,18 +2,18 @@ gqs_script <- function(full_data = FALSE) {
   # data block -----
   data_block <- glue::glue(
     "data {{
-      int<lower=1> I;                 // number of items
-      int<lower=1> R;                 // number of respondents
-      int<lower=1> N;                 // number of observations
-      int<lower=1> C;                 // number of classes
-      int<lower=1> A;                 // number of attributes
-      int<lower=1,upper=I> ii[N];     // item for observation n
-      int<lower=1,upper=R> rr[N];     // respondent for observation n
-      int<lower=0,upper=1> y[N];      // score for observation n
-      int<lower=1,upper=N> start[R];  // starting row for respondent R
-      int<lower=1,upper=I> num[R];    // number of rows (items) for respondent R
-      matrix[C,A] Alpha;              // attribute pattern for each class
-      matrix[I,C] Xi;                 // class attribute mastery indicator
+      int<lower=1> I;                      // number of items
+      int<lower=1> R;                      // number of respondents
+      int<lower=1> N;                      // number of observations
+      int<lower=1> C;                      // number of classes
+      int<lower=1> A;                      // number of attributes
+      array[N] int<lower=1,upper=I> ii;    // item for observation n
+      array[N] int<lower=1,upper=R> rr;    // respondent for observation n
+      array[N] int<lower=0,upper=1> y;     // score for observation n
+      array[R] int<lower=1,upper=N> start; // starting row for respondent R
+      array[R] int<lower=1,upper=I> num;   // number of rows (items) for respondent R
+      matrix[C,A] Alpha;                   // attribute pattern for each class
+      matrix[I,C] Xi;                      // class attribute mastery indicator
     }}"
   )
 
@@ -39,13 +39,13 @@ gqs_script <- function(full_data = FALSE) {
 
   gqs_block <- glue::glue(
     "generated quantities {{",
-    "  matrix[R,C] prob_resp_class;   // post prob of respondent R in class C",
-    "  matrix[R,A] prob_resp_attr;    // post prob of respondent R master A",
-    "  {ifelse(full_data, \"int y_rep[N];\n  int r_class[R];\n\", \"\")}",
+    "  matrix[R,C] prob_resp_class;         // post prob of respondent R in class C",
+    "  matrix[R,A] prob_resp_attr;          // post prob of respondent R master A",
+    "  {ifelse(full_data, \"array[N] int y_rep;\n  array[R] int r_class;\n\", \"\")}",
     "  for (r in 1:R) {{",
     "    row_vector[C] prob_joint;",
     "    for (c in 1:C) {{",
-    "      real log_items[num[r]];",
+    "      array[num[r]] real log_items;",
     "      for (m in 1:num[r]) {{",
     "        int i = ii[start[r] + m - 1];",
     "        log_items[m] = y[start[r] + m - 1] * log(pi[i,c]) +",
@@ -83,18 +83,18 @@ loglik_script <- function() {
   # data block -----
   data_block <- glue::glue(
     "data {{
-      int<lower=1> I;                 // number of items
-      int<lower=1> R;                 // number of respondents
-      int<lower=1> N;                 // number of observations
-      int<lower=1> C;                 // number of classes
-      int<lower=1> A;                 // number of attributes
-      int<lower=1,upper=I> ii[N];     // item for observation n
-      int<lower=1,upper=R> rr[N];     // respondent for observation n
-      int<lower=0,upper=1> y[N];      // score for observation n
-      int<lower=1,upper=N> start[R];  // starting row for respondent R
-      int<lower=1,upper=I> num[R];    // number of rows (items) for respondent R
-      matrix[C,A] Alpha;              // attribute pattern for each class
[…]
