+++ R/stan-lcdm.R
@@ -3,18 +3,18 @@ lcdm_script <- function(qmatrix, prior = NULL, strc = "unconstrained",
   # data block -----
   data_block <- glue::glue(
     "data {{
-      int<lower=1> I;                 // number of items
-      int<lower=1> R;                 // number of respondents
-      int<lower=1> N;                 // number of observations
-      int<lower=1> C;                 // number of classes
-      int<lower=1> A;                 // number of attributes
-      int<lower=1,upper=I> ii[N];     // item for observation n
-      int<lower=1,upper=R> rr[N];     // respondent for observation n
-      int<lower=0,upper=1> y[N];      // score for observation n
-      int<lower=1,upper=N> start[R];  // starting row for respondent R
-      int<lower=1,upper=I> num[R];    // number of rows (items) for respondent R
-      matrix[C,A] Alpha;              // attribute pattern for each class
-      matrix[I,C] Xi;                 // class attribute mastery indicator
+      int<lower=1> I;                      // number of items
+      int<lower=1> R;                      // number of respondents
+      int<lower=1> N;                      // number of observations
+      int<lower=1> C;                      // number of classes
+      int<lower=1> A;                      // number of attributes
+      array[N] int<lower=1,upper=I> ii;    // item for observation n
+      array[N] int<lower=1,upper=R> rr;    // respondent for observation n
+      array[N] int<lower=0,upper=1> y;     // score for observation n
+      array[R] int<lower=1,upper=N> start; // starting row for respondent R
+      array[R] int<lower=1,upper=I> num;   // number of rows (items) for respondent R
+      matrix[C,A] Alpha;                   // attribute pattern for each class
+      matrix[I,C] Xi;                      // class attribute mastery indicator
     }}"
   )
 
@@ -177,15 +177,15 @@ lcdm_script <- function(qmatrix, prior = NULL, strc = "unconstrained",
 
   model_block <- glue::glue(
     "model {{",
-    "  real ps[C];",
     "",
     "  ////////////////////////////////// priors",
     "  {glue::glue_collapse(all_priors, sep = \"\n  \")}",
     "",
     "  ////////////////////////////////// likelihood",
     "  for (r in 1:R) {{",
+    "    row_vector[C] ps;",
     "    for (c in 1:C) {{",
-    "      real log_items[num[r]];",
+    "      array[num[r]] real log_items;",
     "      for (m in 1:num[r]) {{",
     "        int i = ii[start[r] + m - 1];",
     "        log_items[m] = y[start[r] + m - 1] * log(pi[i,c]) +",
