+++ dev/pkgdown.yml
@@ -3,8 +3,9 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   model-estimation: model-estimation.html
+  model-evaluation: model-evaluation.html
   measr: measr.html
-last_built: 2023-06-20T15:35Z
+last_built: 2023-09-20T16:47Z
 urls:
   reference: https://measr.info/reference
   article: https://measr.info/articles
