+++ tests/testthat/_snaps/lcdm-ecpe-code/ecpe.md
@@ -13,7 +13,7 @@
         array[N] int<lower=1,upper=R> rr;    // respondent for observation n
         array[N] int<lower=0,upper=1> y;     // score for observation n
         array[R] int<lower=1,upper=N> start; // starting row for respondent R
-        array[R] int<lower=1,upper=I> num;   // number of rows (items) for respondent R
+        array[R] int<lower=1,upper=I> num;   // number of items for respondent R
         matrix[C,A] Alpha;                   // attribute pattern for each class
         matrix[I,C] Xi;                      // class attribute mastery indicator
       }
