### 2023-09-20 1aa9c47
Add crum functionality to get_parameters
