### 2023-09-20 15acc83
get_parameters preserves item ids by default
