+++ R/utils.R
@@ -54,6 +54,8 @@ profile_labels <- function(attributes) {
 #'   present in the Q-matrix.
 #' @param rename_att Should attribute names from the `qmatrix` be replaced with
 #'   generic, but consistent names (e.g., "att1", "att2", "att3").
+#' @param rename_item Should item names from the `qmatrix` be replaced with
+#'   generic, but consistent names (e.g., 1, 2, 3).
 #'
 #' @return A [tibble][tibble::tibble-package] with one row per parameter.
 #' @export
@@ -64,6 +66,7 @@ profile_labels <- function(attributes) {
 #' get_parameters(ecpe_qmatrix, item_id = "item_id", type = "lcdm",
 #'                rename_att = TRUE)
 get_parameters <- function(qmatrix, item_id = NULL, rename_att = FALSE,
+                           rename_item = FALSE,
                            type = c("lcdm", "dina", "dino", "crum"),
                            attribute_structure = c("unconstrained",
                                                    "independent")) {
@@ -71,6 +74,10 @@ get_parameters <- function(qmatrix, item_id = NULL, rename_att = FALSE,
   qmatrix <- check_qmatrix(qmatrix, identifier = item_id, item_levels = NULL,
                            name = "qmatrix")
   att_names <- colnames(qmatrix)[which(!(colnames(qmatrix) == "item_id"))]
+
+  item_ids <- qmatrix %>%
+    dplyr::select("item_id") %>%
+    tibble::rowid_to_column(var = "item_number")
 
   qmatrix <- qmatrix %>%
     dplyr::select(-"item_id") %>%
@@ -124,6 +131,13 @@ get_parameters <- function(qmatrix, item_id = NULL, rename_att = FALSE,
                                                     att_names[i],
                                                     .data$attributes))
     }
+  }
+
+  if (!rename_item) {
+    all_params <- all_params %>%
+      dplyr::left_join(item_ids, by = c("item_id" = "item_number")) %>%
+      dplyr::mutate(item_id = .data$item_id.y) %>%
+      dplyr::select(-"item_id.y")
   }
 
   strc_params <- if (attribute_structure == "unconstrained") {
