### 2023-09-20 24ddb03
remove old ECPE language from estimation article
