+++ NEWS.md
@@ -4,10 +4,14 @@
 
 * Functions for evaluating estimated models (e.g., `fit_ppmc()`, `reliability()`) no longer recalculate indices if they have previously been saved to the model object. This behavior can be overwritten with `force = TRUE`.
 
+* Updated *Stan* syntax to be compatible with the new array syntax (@andrjohns, #36)
+
+* `get_parameters()` now preserves item identifiers by default. Items can be renamed with numbers (e.g., 1, 2, 3, ...) by setting `rename_item = TRUE`.
+
 * Updates to documentation
   * A new article on model evaluation has been added.
   * The model estimation article has been updated to use the same data set as the model evaluation article.
-  * More detailed installation instructions have been added to the getting started vignette.
+  * More detailed installation instructions have been added to the getting started vignette (#23).
 
 # measr 0.3.1
 
