+++ dev/pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   model-estimation: model-estimation.html
   model-evaluation: model-evaluation.html
   measr: measr.html
-last_built: 2023-09-20T16:47Z
+last_built: 2023-09-20T21:16Z
 urls:
   reference: https://measr.info/reference
   article: https://measr.info/articles
