### 2023-09-20 #5742 @smasongarrison
> Hi @smasongarrison, @Wenchao-Ma 👋 Just checking in to see whether I can be of any support?
>
> It would be helpful if the reviews could be completed in around two weeks, would that work for you? 😊 If you haven't yet generated your checklist, you can do that by simply commenting `@editorialbot generate my checklist` here.

I've finished my review. This paper is very solid on the measurement side of things. I made one tweak to the readme, that should help other readers avoid a pitfall I ran into. The tweaked code will run a model that is *much* faster.
