### 2023-10-02 #5742 @Wenchao-Ma
> > Hi @smasongarrison, @Wenchao-Ma 👋 Just checking in to see whether I can be of any support?
> > It would be helpful if the reviews could be completed in around two weeks, would that work for you? 😊 If you haven't yet generated your checklist, you can do that by simply commenting `@editorialbot generate my checklist` here.
>
> Thanks for your patience @@chartgerink, I've finished my review. This paper is very solid on the measurement side of things. I made one tweak to the readme, that should help other readers avoid a pitfall I ran into. The tweaked code will run a model that is _much_ faster. [smasongarrison/measr#2](https://github.com/smasongarrison/measr/pull/2)
>
> I wish I could comment more on the bayesian side, as I'm not confident in my understanding of the inner workings of STAN to really green lit that piece. (But it _seems_ to be working). Overall, I recommend accepting the paper, contingent on someone with real bayesian/stan expertise checking out that aspect.

I think the paper is well written and very concise. It is an important piece of work in the field. Thank you for inviting me.
