### 2023-10-02 #5742 @chartgerink
@smasongarrison and @Wenchao-Ma thanks a lot for your reviews! I am searching for a Bayes expert before finalizing this round of reviews.

@wjakethompson if you could already tackle the issue highlighted by @smasongarrison (i.e., https://github.com/smasongarrison/measr/pull/2) that'll be helpful to expedite the process.

Thanks everyone for your involvement and energy 💜
