### 2023-10-02 #5742 @seandamiandevine
Thanks for tagging me @laurabustamante. I'm also not familiar with this format of review, but happy to offer my Stan expertise if given a little guidance. @chartgerink, let me know if I can be of assistance.
