### 2023-10-05 #5742 @chartgerink
Thanks @smasongarrison and @laurabustamante for the suggestions!

@seandamiandevine the main difference is that the review focuses on the quality of the software - we provide a checklist for the review here on Github, which helps you make the assessment. At the moment, we are primarily in need of some more eyes on the Bayes + STAN parts, so if you're willing to help you can focus on those bits 😊

Somebody else also is willing to help review so I'll add them now.
