### 2023-10-05 #5742 @seandamiandevine
@chartgerink Thanks for the rundown, Chris. Happy to lend some help. I'll try to get to this as soon as possible.
