### 2023-10-19 9c46dc5
include ppmc example in pkgdown site
