+++ R/ppmc.R
@@ -92,12 +92,12 @@
 #'   \doi{10.35542/osf.io/jzqs8}
 #'
 #' @export
-#' @examplesIf interactive()
+#' @examplesIf measr_examples()
 #' cmds_mdm_dina <- measr_dcm(
 #'   data = mdm_data, missing = NA, qmatrix = mdm_qmatrix,
 #'   resp_id = "respondent", item_id = "item", type = "dina",
 #'   method = "mcmc", seed = 63277, backend = "rstan",
-#'   iter = 700, warmup = 500, chains = 2
+#'   iter = 700, warmup = 500, chains = 2, refresh = 0
 #' )
 #'
 #' fit_ppmc(cmds_mdm_dina, model_fit = "raw_score", item_fit = NULL)
