+++ man/fit_ppmc.Rd
@@ -99,12 +99,12 @@ respondent in each class provides a correct response (\code{item_fit = "conditio
 et al. (2006).
 }
 \examples{
-\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+\dontshow{if (measr_examples()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 cmds_mdm_dina <- measr_dcm(
   data = mdm_data, missing = NA, qmatrix = mdm_qmatrix,
   resp_id = "respondent", item_id = "item", type = "dina",
   method = "mcmc", seed = 63277, backend = "rstan",
-  iter = 700, warmup = 500, chains = 2
+  iter = 700, warmup = 500, chains = 2, refresh = 0
 )
 
 fit_ppmc(cmds_mdm_dina, model_fit = "raw_score", item_fit = NULL)
