### 2023-10-19 55a2a09
start ecpe case study
