+++ vignettes/articles/case-studies/ecpe.Rmd
@@ -1,0 +1,36 @@
+---
+title: "Examination for the Certificate of Proficiency in English"
+output: rmarkdown::html_vignette
+bibliography: bib/references.bib
+csl: bib/apa.csl
+link-citations: true
+vignette: >
+  %\VignetteIndexEntry{Examination for the Certificate of Proficiency in English}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+library(wjake)
+library(showtext)
+
+set_theme(base_family = "Open Sans",
+          plot_margin = ggplot2::margin(10, 10, 10, 10))
+
+font_add_google("Open Sans")
+showtext_auto()
+showtext_opts(dpi = 192)
+
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 7.2916667,
+  fig.align = "center",
+  out.width = "90%"
+)
+
+options(mc.cores = 4,
+        tidyverse.quiet = TRUE)
+```
+
+Some placeholder text...
