+++ vignettes/articles/case-studies/bib/references.bib
@@ -1,0 +1,295 @@
+@incollection{bradshaw-dcm,
+  title = {Diagnostic Classification Models},
+  shorttitle = {Dcm},
+  booktitle = {The {{Handbook}} of {{Cognition}} and {{Assessment}}: {{Frameworks}}, {{Methodologies}}, and {{Applications}}},
+  author = {Bradshaw, Laine},
+  editor = {Rupp, Andr{\'e} A. and Leighton, Jacqueline},
+  year = {2016},
+  edition = {1st},
+  pages = {297--327},
+  publisher = {{John Wiley \& Sons}},
+  address = {{New York, NY}},
+  doi = {10.1002/9781118956588.ch13}
+}
+
+@book{rupp-dcm,
+  title = {Diagnostic Measurement: {{Theory}}, Methods, and Applications},
+  shorttitle = {Diagnostic Measurement},
+  author = {Rupp, Andr{\'e} A. and Templin, Jonathan and Henson, Robert A.},
+  year = {2010},
+  series = {Methodology in the {{Social Sciences}}},
+  publisher = {{Guilford Press}},
+  address = {{New York, NY}}
+}
+
+@article{lcdm,
+  title = {Defining a Family of Cognitive Diagnosis Models Using Log-Linear Models with Latent Variables},
+  author = {Henson, Robert and Templin, Jonathan L. and Willse, John T.},
+  year = {2009},
+  month = jun,
+  journal = {Psychometrika},
+  volume = {74},
+  number = {2},
+  pages = {191--210},
+  issn = {00333123},
+  doi = {10.1007/s11336-008-9089-5}
+}
+
+@article{dina,
+  title = {Higher-order latent trait models for cognitive diagnosis},
+  author = {{de la Torre}, Jimmy and Douglas, Jeffrey A.},
+  year = {2004},
+  journal = {Psychometrika},
+  volume = {69},
+  number = {3},
+  pages = {333--353},
+  issn = {18600980},
+  doi = {10.1007/BF02295640}
+}
+
+@incollection{henson2019,
+  title = {Loglinear Cognitive Diagnostic Model ({{LCDM}})},
+  booktitle = {Handbook of {{Diagnostic Classification Models}}},
+  author = {Henson, Robert and Templin, Jonathan L.},
+  editor = {{von Davier}, Matthias and Lee, Young-Sun},
+  year = {2019},
+  pages = {171--185},
+  publisher = {{Springer International Publishing}},
+  address = {{Cham}},
+  doi = {10.1007/978-3-030-05584-4_8}
+}
+
+@article{chen2018,
+  title = {Applying the ${M}_2$ Statistic to Evaluate the Fit of Diagnostic Classification Models in the Presence of Attribute Hierarchies},
+  author = {Chen, Fu and Liu, Yanlou and Xin, Tao and Cui, Ying},
+  year = {2018},
+  month = oct,
+  journal = {Frontiers in Psychology},
+  volume = {9},
+  issn = {1664-1078},
+  doi = {10.3389/fpsyg.2018.01875}
+}
+
+@incollection{liu2019,
+  title = {Estimating {{CDMs}} Using {{MCMC}}},
+  booktitle = {Handbook of {{Diagnostic Classification Models}}},
+  author = {Liu, Xiang and Johnson, Matthew S.},
+  editor = {{von Davier}, Matthias and Lee, Young-Sun},
+  year = {2019},
+  pages = {629--646},
+  publisher = {{Springer International Publishing}},
+  address = {{Cham}},
+  doi = {10.1007/978-3-030-05584-4_31}
+}
+
+@article{hdcm,
[…]
