### 2023-10-24 wjakethompson/measr#38 @seandamiandevine
JOSS Review
Hi! @chartgerink asked me to[ review this package for JOSS ](https://github.com/openjournals/joss-reviews/issues/5742)and focus particularly on the Stan implementation. I am not a content expert on DCMs or IRT, so I will focus on the Bayesian modeling side. Overall, code all seems appropriate. Below are some thoughts and questions I had. Some of these may be non-issues and stem from my ignorance of common practice in the field.

- In the DINA model code, why is each slip and guess parameter provided its own prior? Can the user specify priors for each individual item? In the `measrprior()` function, only a single specification for guess and slip parameters can be provided, so it may not be necessary to specify these separately in the Stan model. Unless there is some benefit to doing so, consider vectorizing the parameter list.
- In the DINA model code, what is the motivation behind a B(5,25) prior for the slip and guess parameters. This puts the prior mode of each parameter at roughly 0.14. Is this plausible for users who may be working with smaller pilot sets and who keep default priors? This is particularly important in light of my previous point.
- In the LCDM model code, why are interactions bounded at the minimum of the main effects?
- For all likelihoods in the Stan models, the many nested for loops may slow things down significantly. Consider reorganizing the Stan data into a R x C matrix to allow for vectorized operations. This will speed things up and avoid the need for nested looping. Note however, this would represent a significant refactor, so take this is a suggestion to improve performance rather than a prerequisite for publication.

Open issues:

- [ ] Vectorize priors in DINA model
- [ ] Double-check DINA priors
- [ ] Check constraints on interactions effects in LCDM
- [ ] Vectorize likelihood functions?
