### 2023-10-24 #5742 @seandamiandevine
Thanks so much @chartgerink for the patience. I have finished my review. Overall, the code all seems appropriate. I am not a content expert on DCMs or IRT, so I focused on the Bayesian modeling side as requested.

I opened an [issue](https://github.com/wjakethompson/measr/issues/38) on the original repo detailing my recommendations.

Also, as a small note on the earlier thread, I would be careful with @smasongarrison's suggested patch (https://github.com/smasongarrison/measr/pull/2), because while faster it only estimates the mode of the Bayesian posterior using Stan's optimizer, which may go against the "spirit" of the package for newcomers consulting the README. Just a thought.

I think the paper is well written and clear. The package is well-organized and efficient. Thank you for inviting me as a reviewer!
