### 2023-10-24 7afcbf2
start ecpe case study
