+++ vignettes/articles/case-studies/bib/references.bib
@@ -1,3 +1,4 @@
+% General DCM references -------------------------------------------------------
 @incollection{bradshaw-dcm,
   title = {Diagnostic Classification Models},
   shorttitle = {Dcm},
@@ -22,9 +23,11 @@
   address = {{New York, NY}}
 }
 
+
+% Specific models --------------------------------------------------------------
 @article{lcdm,
   title = {Defining a Family of Cognitive Diagnosis Models Using Log-Linear Models with Latent Variables},
-  author = {Henson, Robert and Templin, Jonathan L. and Willse, John T.},
+  author = {Henson, Robert and Templin, Jonathan and Willse, John T.},
   year = {2009},
   month = jun,
   journal = {Psychometrika},
@@ -33,6 +36,29 @@
   pages = {191--210},
   issn = {00333123},
   doi = {10.1007/s11336-008-9089-5}
+}
+
+@incollection{lcdm-handbook,
+  title = {Loglinear Cognitive Diagnostic Model ({{LCDM}})},
+  booktitle = {Handbook of {{Diagnostic Classification Models}}},
+  author = {Henson, Robert and Templin, Jonathan},
+  editor = {{von Davier}, Matthias and Lee, Young-Sun},
+  year = {2019},
+  pages = {171--185},
+  publisher = {{Springer International Publishing}},
+  address = {{Cham}},
+  doi = {10.1007/978-3-030-05584-4_8}
+}
+
+@article{hdcm,
+  title = {Hierarchical Diagnostic Classification Models: {{A}} Family of Models for Estimating and Testing Attribute Hierarchies},
+  author = {Templin, Jonathan and Bradshaw, Laine},
+  year = {2014},
+  journal = {Psychometrika},
+  volume = {79},
+  number = {2},
+  pages = {317--339},
+  doi = {10.1007/s11336-013-9362-0}
 }
 
 @article{dina,
@@ -47,18 +73,8 @@
   doi = {10.1007/BF02295640}
 }
 
-@incollection{henson2019,
-  title = {Loglinear Cognitive Diagnostic Model ({{LCDM}})},
-  booktitle = {Handbook of {{Diagnostic Classification Models}}},
-  author = {Henson, Robert and Templin, Jonathan L.},
-  editor = {{von Davier}, Matthias and Lee, Young-Sun},
-  year = {2019},
-  pages = {171--185},
-  publisher = {{Springer International Publishing}},
-  address = {{Cham}},
-  doi = {10.1007/978-3-030-05584-4_8}
-}
 
+% Absolute model fit -----------------------------------------------------------
 @article{chen2018,
   title = {Applying the ${M}_2$ Statistic to Evaluate the Fit of Diagnostic Classification Models in the Presence of Attribute Hierarchies},
   author = {Chen, Fu and Liu, Yanlou and Xin, Tao and Cui, Ying},
@@ -68,94 +84,6 @@
   volume = {9},
   issn = {1664-1078},
   doi = {10.3389/fpsyg.2018.01875}
-}
-
-@incollection{liu2019,
-  title = {Estimating {{CDMs}} Using {{MCMC}}},
-  booktitle = {Handbook of {{Diagnostic Classification Models}}},
-  author = {Liu, Xiang and Johnson, Matthew S.},
-  editor = {{von Davier}, Matthias and Lee, Young-Sun},
-  year = {2019},
-  pages = {629--646},
-  publisher = {{Springer International Publishing}},
-  address = {{Cham}},
-  doi = {10.1007/978-3-030-05584-4_31}
-}
-
-@article{hdcm,
-  title = {Hierarchical Diagnostic Classification Models: {{A}} Family of Models for Estimating and Testing Attribute Hierarchies},
[…]
