+++ vignettes/articles/case-studies/ecpe.Rmd
@@ -2,7 +2,7 @@
 title: "Examination for the Certificate of Proficiency in English"
 output: rmarkdown::html_vignette
 bibliography: bib/references.bib
-csl: bib/apa.csl
+csl: ../../bib/apa.csl
 link-citations: true
 vignette: >
   %\VignetteIndexEntry{Examination for the Certificate of Proficiency in English}
@@ -33,4 +33,224 @@ options(mc.cores = 4,
         tidyverse.quiet = TRUE)
 ```
 
-Some placeholder text...
+The *Examination for the Certificate of Proficiency in English* (ECPE) is an assessment that measures advanced English skills for individuals for whom English is not the primary language.
+In this case study, we'll use data from the grammar section of the ECPE, which measures 3 skills with 28 items: morphosyntactic rules, cohesive rules, and lexical rules.
+This data set has previously been used by @templin-emip-2013 and @mcmc-handbook to demonstrate how to estimate diagnostic classification models (DCMs) with Mplus and Markov chain Monte Carlo (MCMC), respectively.
+Additionally, @hdcm used this ECPE data as a motivating example for developing a hierarchical DCM, and @chen2018 used the ECPE data to evaluate the effectiveness of the *M<sub>2</sub>* statistic for assessing model fit in the presence of attribute hierarchies.
+
+
+## Explore the Data
+
+The ECPE data is built into measr and can be accessed by loading the package.
+A complete description of the data can be viewed using `?ecpe_data`.
+
+```{r view-data}
+library(measr)
+
+ecpe_data
+```
+
+We can see that the data set has one row for each respondent, and that `r nrow(ecpe_data)` respondents completed this section of the ECPE.
+We also see that the data has `r ncol(ecpe_data)` columns.
+The first column contains respondent identifiers, and the remaining `r ncol(ecpe_data) - 1` columns contain dichotomous item responses for each items.
+The item responses are coded as 0 for an incorrect response and 1 for a correct response.
+
+In addition to the data, we also have a Q-matrix that define which attributes are measured by each item.
+The Q-matrix has `r nrow(ecpe_qmatrix)` rows, which corresponds to the total number of items.
+The first column of the Q-matrix contains item identifiers, which are the same as the column names in `ecpe_data` that contain item responses.
+The remaining columns define the attributes measured by the ECPE.
+A value of 0 indicates that the item does not measure that attribute, whereas a value of 1 indicates that the attribute is measured by that item.
+For example, item E1 measures both morphosyntactic rules and cohesive rules, and item E4 measures only lexical rules.
+
+```{r view-qmatrix}
+ecpe_qmatrix
+```
+
+For a quick summary of the data, we can calculate the proportion of respondents that answered each question correctly (i.e., the item *p*-values).
+
+```{r pvalues}
+library(tidyverse)
+
+ecpe_data %>% 
+  summarize(across(-resp_id, mean)) %>% 
+  pivot_longer(everything(), names_to = "item_id", values_to = "pvalue")
+```
+
+We can then join the item *p*-values with the Q-matrix to get a sense of which attributes are the most difficult.
+Overall, most of the items have relatively high *p*-values, with most items having a *p*-value between .6 and .9.
+Note that in general, items measuring morphosyntactic rules tend to be the most difficult (i.e., lower *p*-values), followed by items measuring cohesive rules, and finally items measuring lexical rules.
+
+<details><summary>Plot code</summary>
+
+```{r pvalue-plot, eval = FALSE}
+ecpe_data %>% 
+  summarize(across(-resp_id, mean)) %>% 
+  pivot_longer(everything(), names_to = "item_id", values_to = "pvalue") %>% 
[…]
