### 2023-10-24 f58c693
add example of prior for a single item
