### 2023-10-24 #5742 @wjakethompson
Thanks to all! I believe I have responded to all of @seandamiandevine's comments in wjakethompson/measr#38, and will close that issue once Sean approves.

Once that is closed, I believe the only outstanding issue remaining is the recommended changes to the README from @smasongarrison. My preference would be to not make those changes. The change would make the example in the README render much more quickly, but as Sean noted above, that goes against primary goal of the package, which is to make a fully Bayesian estimation of these models more accessible to users.
