### 2023-10-24 #5742 @smasongarrison
I understand that the purpose of the package is to make a fully Bayesian
estimation of these models more accessible to users. So it would seem to me
that ensuring that your users can run the model is key. My experience was
such that the original readme code took so long to run that I couldn't tell
if it was working at all. I strongly recommend implementing a change to
your readme so that the user's initial experience isn't concern or
confusion, after waiting for more than 20 minutes. Frankly, that initial
bottleneck in the readme was a major contributor to why my review took so
long to complete. I doubt that I'm the only one who will have that
less-than-ideal user experience.

S. Mason Garrison
Assistant Professor
Quantitative Psychology

Wake Forest University
(she/they)
<https://mailtrack.io/trace/link/154e75d482620854df48fa66d805dafe0d64fa77?url=https%3A%2F%2Fwww.mypronouns.org%2Fwhat-and-why&userId=7103102&signature=d762eab3c7de303e>
Meet with me (calendly)
<https://mailtrack.io/trace/link/61ccef59cba31588da07f33e09bce18c2090a6bd?url=https%3A%2F%2Fcalendly.com%2Fsmasongarrison%2F&userId=7103102&signature=1ceb2f287306e41c>

Spare computers? Help fight COVID-19
<https://mailtrack.io/trace/link/5393e1b2684056beb59ab9bdae19a79774d7e772?url=https%3A%2F%2Ffoldingathome.org%2Fcategory%2Finfectious-diseases%2Fcovid-19%2F&userId=7103102&signature=d77620015e9d450a>

On Tue, Oct 24, 2023 at 5:07 PM Jake Thompson ***@***.***>
wrote:

> Thanks to all! I believe I have responded to all of @seandamiandevine
> <https://github.com/seandamiandevine>'s comments in wjakethompson/measr#38
> <https://github.com/wjakethompson/measr/issues/38>, and will close that
> issue once Sean approves.
>
> Once that is closed, I believe the only outstanding issue remaining is the
> recommended changes to the README from @smasongarrison
> <https://github.com/smasongarrison>. My preference would be to not make
> those changes. The change would make the example in the README render much
> more quickly, but as Sean noted above, that goes against primary goal of
> the package, which is to make a fully Bayesian estimation of these models
> more accessible to users.
>
