### 2023-10-24 wjakethompson/measr#38 @seandamiandevine
Thanks Jake. This makes sense to me. I have some thoughts about vectorizing, but I will think about them more deeply and open another issue (outside of the scope of this review) if need be.

Great job and great package!
