### 2023-10-24 #5742 @seandamiandevine
I am satisfied with @wjakethompson's responses and have closed the [issue](https://github.com/wjakethompson/measr/issues/38) on the original repo. Great job on a great package!
