### 2023-10-25 dd77337
update readme to include note about computation time
