### 2023-10-25 #5742 @wjakethompson
@smasongarrison I've made some proposed changes to the README here: wjakethompson/measr#40. I basically flipped your proposed changes. That is, rather than changing the estimation and noting in the text that MCMC could be alternatively specified, I kept the estimation the same, but noted that it is time intensive and that the optimizer could be used to speed things up. Please take a look and let me know if that help alleviate your concerns, and if so, I will merge those changes in.
