### 2023-10-26 #5742 @chartgerink
Perfect - thank you to all the reviewers and @wjakethompson for handling everything promptly 😊

Looking over the positive reviews, I see that only a "Contributing" section may still be of help for the README. This to help people understand where they can go with issues, and whether you offer any support. Beyond all of that, it seems like the evaluation is very positive - great job!

Once some form of contributor info is added, I'm happy to accept the paper and move on to final checks + request archival of the software (e.g., in Zenodo). Thank you so much for the ongoing patience for this paper.
