### 2023-10-26 #5742 @wjakethompson
@chartgerink Thanks! Contributor info has now been added (wjakethompson/measr#41).
