+++ dev/pkgdown.yml
@@ -3,9 +3,9 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   model-estimation: model-estimation.html
-  model-evaluation: model-evaluation.html
   measr: measr.html
-last_built: 2023-09-20T21:16Z
+  paper: paper.html
+last_built: 2023-10-31T13:40Z
 urls:
   reference: https://measr.info/reference
   article: https://measr.info/articles
