+++ dev/pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   model-estimation: model-estimation.html
   measr: measr.html
   paper: paper.html
-last_built: 2023-10-31T13:40Z
+last_built: 2023-10-31T13:58Z
 urls:
   reference: https://measr.info/reference
   article: https://measr.info/articles
