### 2023-10-31 #5742 @wjakethompson
Great! Here's the tag and DOI information. Let me know if there is anything else you need from me!

Tag: [v0.3.1-joss](https://github.com/wjakethompson/measr/releases/tag/v0.3.1-joss)
DOI: 10.5281/zenodo.10058365
