### 2023-10-31 #5742 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
