### 2023-11-06 #5742 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
